+++ src/main/webapp/WEB-INF/sql/equation_table.sql
@@ -5,11 +5,11 @@ DROP TABLE IF EXISTS `equation`;
 
 CREATE TABLE `equation`
 (
-    `id`          int(11) NOT NULL AUTO_INCREMENT,
-    `a`           int(11) DEFAULT NULL,
-    `b`           int(11) DEFAULT NULL,
-    `c`           int(11) DEFAULT NULL,
-    `x1`           int(11) DEFAULT NULL,
-    `x2`           int(11) DEFAULT NULL,
+    `id`          int  NOT NULL AUTO_INCREMENT,
+    `a`           DOUBLE         DEFAULT NULL,
+    `b`           DOUBLE        DEFAULT NULL,
+    `c`           DOUBLE      DEFAULT NULL,
+    `x1`          DOUBLE      DEFAULT NULL,
+    `x2`          DOUBLE     DEFAULT NULL,
     PRIMARY KEY (`id`)
 ) ENGINE = InnoDB AUTO_INCREMENT = 1 DEFAULT CHARSET = latin1;
